requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.3.1
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
openpyxl==3.1.5
xlsxwriter==3.2.0
lxml==5.2.2
reportlab==4.2.2
gunicorn==22.0.0
werkzeug==3.0.3
